+++ Assets/Plugins/Appboy/models/InAppMessage/InAppMessageBase.cs
@@ -7,9 +7,6 @@ using System.Text;
 using UnityEngine;
 
 namespace Appboy.Models.InAppMessage {
-
-  // Use InAppMessageFactory.BuildInAppMessage(InAppMessageString) to get the in-app message with correct type from
-  // Appboy SDK.
   public abstract class InAppMessageBase : IInAppMessage {
     private const int DefaultDuration = 5000;
     protected string _jsonString;
@@ -17,49 +14,33 @@ namespace Appboy.Models.InAppMessage {
     private bool _impressionLogged = false;
     private int _duration;
 
-    // BackgroundColor defines the background color of the in-app message.
+    // The background color of the in-app message.
     public Color? BackgroundColor { get; set; }
 
-    // TextColor defines the message text color of the in-app message.
     public Color? TextColor { get; set; }
 
-    // Icon defines the font awesome unicode string of the Appboy icon.
-    // You can choose to display one of the Appboy icons from Appboy dashboard. When you do so, this property will have the
-    // unicode string of font awesome.
+    // The font awesome unicode string of the in-app message icon.
     public string Icon { get; set; }
 
-    // IconColor defines the font color of icon property.
+    // The color of the icon text.
     public Color? IconColor { get; set; }
 
-    // IconBackgroundColor defines the background color of icon property.
+    // The color of the icon background.
     public Color? IconBackgroundColor { get; set; }
 
-    // imageURI defines the URI of the image icon on in-app message.
-    // When there is a iconImage defined, the iconImage will be used and the value of property icon will 
-    // be ignored.
     public string ImageURI { get; set; }
 
-    // This property defines the message displayed within the in-app message.
     public string Message { get; set; }
 
-    // This property carries extra data in the form of an dictionary which can be sent down via the Appboy Dashboard.
-    // You may want to design and implement a custom handler to access this data depending on your use-case.
     public Dictionary<string, string> Extras { get; set; }
 
-    // This property defines the action that will be performed when the in-app message is clicked.
-    // See the ClickAction enum documentation above offers additional details.
     public ClickAction InAppClickAction { get; private set; }
 
-    // When the in-app message's InAppClickAction is URI, clicking on the in-app message will redirect to the uri defined
-    // in this property.
-    //
-    // This property can be a HTTP URI or a protocol URI.
     public string URI { get; private set; }
 
-    // InAppDismissType defines the dismissal behavior of the in-app message.
     public DismissType InAppDismissType { get; set; }
 
-    // This property defines the number of seconds before the in-app message is automatically dismissed.
+    // The number of seconds that should elapse before the in-app message is automatically dismissed.
     public int Duration { 
       get {
         return _duration;
@@ -81,7 +62,7 @@ namespace Appboy.Models.InAppMessage {
     
     public InAppMessageBase(JSONClass json) {
       if (json == null) {
-        throw new ArgumentNullException("The JSON Class passed to InAppMessage constructor is null.");
+        throw new ArgumentNullException("Received null JSONClass.");
       }
       _jsonString = json.ToString();
       Message = json[InAppMessageConstants.MessageKey];
@@ -92,7 +73,7 @@ namespace Appboy.Models.InAppMessage {
       URI = json[InAppMessageConstants.URIKey];
       ImageURI = json[InAppMessageConstants.ImageURLKey];
       if (InAppClickAction == ClickAction.URI && URI == null) {
-        Debug.Log("The click action cannot be set to URI because the uri is null. Setting click action to NONE.");
+        Debug.Log("Received ClickAction.URI but URI was null. Setting click action to NONE.");
         InAppClickAction = ClickAction.NONE;
       }
       InAppDismissType = (DismissType)EnumUtils.TryParse(typeof(DismissType), json[InAppMessageConstants.DismissTypeKey], true, DismissType.AUTO_DISMISS);
@@ -109,7 +90,7 @@ namespace Appboy.Models.InAppMessage {
         _clickLogged = true;
         AppboyBinding.LogInAppMessageClicked(_jsonString);
       } else {
-        Debug.Log("The in-app message already logged a click.");
+        Debug.Log("Already logged a click. Doing nothing.");
       }
     }
     
@@ -118,7 +99,7 @@ namespace Appboy.Models.InAppMessage {
         _impressionLogged = true;
         AppboyBinding.LogInAppMessageImpression(_jsonString);
       } else {
-        Debug.Log("The in-app message already logged an impression.");
+        Debug.Log("Already logged an impression. Doing nothing.");
       }
     }
 
@@ -129,7 +110,7 @@ namespace Appboy.Models.InAppMessage {
         URI = null;
         return true;
       } else {
-        Debug.LogError("A non-null URI is required in order to set the InAppClickAction to URI.");
+        Debug.Log("A non-null URI is required in order to set the InAppClickAction to URI.");
         return false;
       }
     }