+++ Assets/Plugins/Appboy/models/InAppMessage/InAppMessageFactory.cs
@@ -2,13 +2,23 @@ using System;
 using Appboy.Utilities;
 
 namespace Appboy.Models.InAppMessage {
-  // This class is the in-app message builder. Passing the in-app message JSON string from Appboy, and the Builder will
-  // return a corresponding InAppMessageSlideup, InAppMessageModal or InAppMessageFull;
   public class InAppMessageFactory {
+
+    /// <summary>
+    /// Builds in-app message instances from serialized in-app message objects.
+    ///
+    /// Use this method to create a useful model object from the in-app message string
+    /// passed into your gameobject callback.
+    /// </summary>
+    /// <param name="inAppMessageJSONString">
+    /// The serialized in-app message object.
+    /// </param>
+    ///<returns>
+    /// An IInAppMessage instance populated by the serialized in-app message object.
+    /// </returns>
     public static IInAppMessage BuildInAppMessage(string inAppMessageJSONString) {
       JSONClass json = InAppMessageConstants.JSONObjectFromString(inAppMessageJSONString);
       if (json != null) {
-        // If the type is specified, create corresponding in-app message, otherwise, create an in-app message slideup.
         InAppMessageType type = (InAppMessageType)EnumUtils.TryParse(typeof(InAppMessageType), json[InAppMessageConstants.TypeKey], true, InAppMessageType.SLIDEUP);
         switch (type) {
           case InAppMessageType.FULL: