+++ unity-samples/Assets/Editor/SwitchPlatform.cs
@@ -6,10 +6,10 @@ using System;
 public class AppboySwitchPlatform : MonoBehaviour {
 
   static void SwitchToAndroid () {
-    EditorUserBuildSettings.SwitchActiveBuildTarget (BuildTarget.Android);
+    EditorUserBuildSettings.SwitchActiveBuildTarget (BuildTargetGroup.Android, BuildTarget.Android);
   }
 
   static void SwitchToIos () {
-    EditorUserBuildSettings.SwitchActiveBuildTarget (BuildTarget.iOS);
+    EditorUserBuildSettings.SwitchActiveBuildTarget (BuildTargetGroup.iOS, BuildTarget.iOS);
   }
 }